add data checking in review API

+++ src/review/review.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 import { CreateReviewDTO } from './review.dto';
 
@@ -7,6 +7,12 @@ export class ReviewService {
   constructor(private prisma: PrismaClient) {}
 
   async createReview(userId: number, data: CreateReviewDTO) {
+    const book = await this.prisma.book.findUnique({
+      where: { id: data.bookId },
+    });
+
+    if (!book) throw new NotFoundException('no book found');
+
     return await this.prisma.review.create({
       data: { bookId: data.bookId, content: data.content, userId },
       select: { id: true },
@@ -14,6 +20,10 @@ export class ReviewService {
   }
 
   async getReviewsByBookId(bookId: number) {
+    const book = await this.prisma.book.findUnique({ where: { id: bookId } });
+
+    if (!book) throw new NotFoundException('no book found');
+
     return await this.prisma.review.findMany({
       where: { bookId },
       omit: { bookId: true },
@@ -21,6 +31,10 @@ export class ReviewService {
   }
 
   async editReview(id: number, data: Partial<CreateReviewDTO>) {
+    const review = await this.prisma.review.findUnique({ where: { id } });
+
+    if (!review) throw new NotFoundException('no review to be updated');
+
     return await this.prisma.review.update({
       where: { id },
       select: { id: true },
@@ -29,6 +43,10 @@ export class ReviewService {
   }
 
   async deleteReview(id: number) {
+    const review = await this.prisma.review.findUnique({ where: { id } });
+
+    if (!review) throw new NotFoundException('no review to be deleted');
+
     return await this.prisma.review.delete({
       where: { id },
       select: { id: true },
